refactored singleproductratings out of singleproductview
