+++ client/components/SingleProductView.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import {Link} from 'react-router-dom'
-import {Row, Col, Carousel, Button, FormGroup, Checkbox, Grid, ListGroup, ListGroupItem} from 'react-bootstrap'
+import {Row, Col, Carousel, Grid} from 'react-bootstrap'
 import UserAddProductForm from './UserAddProductForm.jsx'
 import SingleProductReviews from './SingleProductReviews.jsx'
+import SingleProductRatings from './SingleProductRatings.jsx'
 
 // TODO: Refactor the customer reviews, star ratings, and related products code into their components
 
@@ -12,64 +12,6 @@ export function SingleProductView(props) {
 	const {products, categories} = props
 	const currentProduct = products.filter((product) => product.id === productId)[0]
 	console.log('curr product', currentProduct)
-
-	// Get average review rating
-	const calcAverageRating = () => {
-		return currentProduct.reviews
-			.map((review) => review.rating)
-			.reduce((acc, curr) => acc + curr) / currentProduct.reviews.length
-	}
-	// Convert the average review rating to a .5 interval rating so it can be represented by stars
-	const calcStarRating = (rating) => {
-		const floor = Math.floor(rating)
-		const ceil = Math.ceil(rating)
-		const midPoint = (floor + ceil) / 2
-		const difference = rating - midPoint
-		if (Math.abs(difference) <= 0.25) {
-			return midPoint
-		} else if(difference < 0.25) {
-			return floor
-		} else {
-			return ceil
-		}
-	}
-	// render the star rating
-	const renderStarRating = (starRating) => {
-		var starRatingArr = []
-		if (starRating === Math.round(starRating)) {
-			console.log('equal to rounded')
-			for (var i = 0; i < starRating; i++) {
-				starRatingArr.push('full')
-			}
-			var j = 0
-			return starRatingArr.map(() => {
-				return (
-					<a key={j++} href="#"><i className="fa fa-star"></i></a>
-				)
-			})
-		} else {
-			console.log('starrating not equal to rounded', starRating, Math.round(starRating))
-			for (var i = 0; i < starRating - 1; i++) {
-				console.log(i)
-				starRatingArr.push('full')
-			}
-			starRatingArr.push('half')
-
-			return starRatingArr.map((type) => {
-            	if (type === 'full') {
-					return (
-						<a key={j++} href="#"><i className="fa fa-star"></i></a>
-					)
-				} else {
-					return (
-						<a key={j++} href="#"><i className="fa fa-star-half-o"></i></a>
-					)
-				}
-			})
-		}
-	}
-
-	const starRating = calcStarRating(calcAverageRating())
 
 	if(currentProduct) {
 		return (
@@ -95,11 +37,7 @@ export function SingleProductView(props) {
 					<Col md={6}>
 						<div>
 							<h2>{currentProduct.name}</h2>
-							{
-								renderStarRating(starRating)
-							}
-							<br />
-							<a href="#">{currentProduct.reviews.length} customer reviews</a>
+								<SingleProductRatings currentProduct={currentProduct}/>
 							<hr />
 						List Price: <p style={{color: 'green', display: 'inline-block'}}>${currentProduct.price}</p>
 							<p>Description:</p>
